requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.6.0
SQLAlchemy==1.4.47
markdown2==2.4.0
python-dotenv==0.19.0
